api日志完成

+++ zeus-common-api/src/main/java/jp/tokyo/leon/zeus/common/log/ZeusApiLogEntity.java
@@ -21,7 +21,7 @@ public class ZeusApiLogEntity {
     /**
      * 消耗时间
      */
-    private Integer spendTime;
+    private Long spendTime;
 
     /**
      * 根路径
